Rename wrapped
