+++ examples/nextjs/src/simple-greeter-react.ts
